// frontend/src/modules/content/types.ts
import { UploadFile } from '@/types/file';

import { Company } from '../companies';

import {
  CONTENT_ITEM_TYPE,
  FILES_EVIDENCE_STATE,
  IMAGE_TYPE,
} from './constants';

export type VehicleDetailParams = {
  id: string;
};

export type ContentItemType = keyof typeof CONTENT_ITEM_TYPE;

export type PhotoType = Omit<ContentItemType, 'VIDEO'>;

export type ImageType = keyof typeof IMAGE_TYPE;

type User = {
  id: string;
  firstname: string;
  lastname: string;
  email: string;
  photoUrl: string;
};

export type Vehicles<T> = {
  vehicles: T;
  photoBoxStatus: string;
};

export type VehicleCompany = {
  companyName: string;
};

export type QualityCheckStatus =
  | 'IN_PROGRESS'
  | 'CHECKED'
  | 'CHECKED_WITH_ERRORS'
  | 'UNCHECKED';

export type Vehicle = {
  allImagesArchiveUri: string;
  bodyType: string;
  company: VehicleCompany;
  createdAt: Date;
  fuelType: string;
  id: string;
  make: string;
  model: string;
  modelYear: string;
  processedImagesArchiveSize: number;
  processedImagesArchiveUri: string;
  user: User;
  vin: string;
  photoQualityCheckStatus: QualityCheckStatus;
  photoQualityCheckerUserId: string;
};

export type ImageCount = {
  photoType: ContentItemType;
  count: number;
};

type VehicleContentItem = {
  id: string;
  vehicleId?: string;
  position: number;
  user?: User;
};

export type VehicleImage = VehicleContentItem & {
  image: UploadFile;
  originalImage: UploadFile;
  photoType?: ContentItemType;
  isAutoGenerated: boolean;
};

export type VehicleVideo = VehicleContentItem & {
  video?: UploadFile & { attachmentType: any };
};

export type GeoLocation = {
  latitude: number;
  longitude: number;
  address: string;
};

export type VehicleDetails = {
  coverImage: VehicleImage;
  vehicle: Vehicle;
  imageCounts: ImageCount[];
  images: VehicleImage[];
  videos: VehicleVideo[];
  videosCount: number;
  geoLocation: GeoLocation;
};

export type QualityCheck = {
  isQualityGood: boolean;
  qualityCheckStatus: QualityCheckStatus;
  issues: [];
  comments: string;
};

export type GalleryPhoto = { internalPosition: number } & VehicleImage;

export type GalleryItem = VehicleContentItem & {
  internalPosition: number;
  contentType?: ContentItemType;
  video?: UploadFile;
  qualityCheck?: QualityCheck;
} & Partial<Pick<VehicleImage, 'image' | 'originalImage' | 'isAutoGenerated'>>;

export type FilterValues = {
  bodyTypes: string | null;
  dateFrom: Date | string | null;
  dateTo: Date | string | null;
  fuelTypes: string | null;
  makes: string | null;
  modelYears: string | null;
  models: string | null;
  tenantIds: string[];
  parentTenantIds: string[];
  userIds: string[];
  vIN: string;
};

export type ChangeVinNumberFormValues = {
  vin: string;
};

export type Photographer = {
  userId: string;
  userFirstName: string;
  userLastName: string;
  userEmail: string;
  vehiclesActiveCount: number;
  vehiclesDeletedCount: number;
  vehiclesImagesCount: number;
  totalVehiclesCount: number;
};

export type CardsStatistics = {
  activeVehiclesCount: number;
  deletedVehiclesCount: number;
  totalVehiclesCount: number;
  vehiclesImagesCount: number;
};

export type StatisticsDateFilters = {
  dateFrom?: Date;
  dateTo?: Date;
};

export type StatisticsDateFiltersForApi = {
  dateFrom: string | null;
  dateTo: string | null;
};

export type FileEvidenceState = keyof typeof FILES_EVIDENCE_STATE;

export type CompanyGroup = Pick<Company, 'companyName' | 'tenantId'>;

export type QualityCheckItem = {
  label: string;
  value: string;
  checked: boolean;
};

export type QualityFeedbackProps = {
  onValidatePicture: () => void;
  resetFlag: boolean;
  isModalOpen: boolean;
  vehicleItem: GalleryItem;
};